% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{makehm}
\alias{makehm}
\title{Title}
\usage{
makehm(
  cellchatobj = cellchatobj,
  col_fun = cols,
  sources = c(),
  targets = c(),
  fontsize = 4,
  hmtitle = "title"
)
}
\arguments{
\item{cellchatobj}{a cellchat object}

\item{col_fun}{color function}

\item{sources}{a list of cell types to be used as sender cells}

\item{targets}{a list of cell types to be used as receiver cells}

\item{fontsize}{a number}

\item{hmtitle}{a string for a title}
}
\value{
a list of heatmaps
}
\description{
Title
}
\examples{

hm <- makehm(cellchatobj = cellchatobj, col_fun = cols)
hm[[1]]
}
