% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_relativeheatmap.R
\name{makerelativehm}
\alias{makerelativehm}
\title{This function makes a heatmap with relative pathway singnaling strength}
\usage{
makerelativehm(cellchatobj, col_fun, sources, targets, fontsize, hmtitle)
}
\arguments{
\item{cellchatobj}{a cellchat object}

\item{col_fun}{color function}

\item{sources}{list of cell types that are sender cells}

\item{targets}{list of cell types that are receiver cells}

\item{fontsize}{numeric}

\item{hmtitle}{string as title}
}
\value{
a heatmap in which row (pathway) values are scaled between 0 and 1.
}
\description{
This function makes a heatmap with relative pathway singnaling strength
}
