% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{prephm}
\alias{prephm}
\title{Title}
\usage{
prephm(df.netPx1outgoing)
}
\arguments{
\item{df.netPx1outgoing}{df}
}
\value{
a df for input to the heatmap
}
\description{
Title
}
\examples{
df <- prephm(df)
}
