% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_relativeheatmap.R
\name{preprelativehm}
\alias{preprelativehm}
\title{Preps df for relative heatmap}
\usage{
preprelativehm(df.netPx1outgoing)
}
\arguments{
\item{df.netPx1outgoing}{df}
}
\value{
df for heatmap
}
\description{
Preps df for relative heatmap
}
